테스트 코드 작성하기

+++ app/main.py
@@ -1,18 +1,20 @@
-# app/main.py
+import os
+
 from fastapi import FastAPI
 
-#
 from app.api.endpoints import router as api_router
-from app.api.endpoints import router as auth_router
-
-# from app.auth.auth import router as auth_router
-from app.database import Base, engine
+from app.database import Base, engine, get_db
 
 app = FastAPI()
 
 app.include_router(api_router, prefix="/api", tags=["api"])
 
-# 기존 테이블 삭제
-Base.metadata.drop_all(bind=engine)
-# 새로운 테이블 생성
-Base.metadata.create_all(bind=engine)
+env = os.getenv("ENV", "production")
+
+if env == "development":
+    print("Dropping all tables...")
+    Base.metadata.drop_all(bind=engine)
+    print("Creating all tables...")
+    Base.metadata.create_all(bind=engine)
+else:
+    print("Running in production mode; no automatic table creation or deletion.")
